refactor: Refactor Chat into a page and component
Move API client initialization outside of the component.
