+++ packages/components/src/components/chat/UserMessageHandler.ts
@@ -1,0 +1,3 @@
+export type AckCallback = (messageId: string, threadId: string) => void;
+
+export type UserMessageHandler = (message: string, ack: AckCallback) => void;
